commit 9

+++ posts/signals.py
@@ -5,11 +5,18 @@ from .models import PostMedia, Post
 
 post_delete_flag = False
 
-@receiver(pre_delete, sender=PostMedia)
+@receiver(post_delete, sender=PostMedia)
 def delete_post_media_file(sender, instance, **kwargs):
-    if not post_delete_flag:
-        if instance.file:
+    if not post_delete_flag and instance.file:
             instance.file.delete()
+
+
+@receiver(pre_delete, sender=Post)
+def media_dir(sender, instance, **kwargs):
+    instance._media_files = list(instance.postmedia_set.all())
+
+    if instance._media_files:
+        instance._media_dir = os.path.dirname(instance._media_files[0].file.path)
 
 
 @receiver(post_delete, sender=Post)
@@ -17,15 +24,10 @@ def delete_post_media(sender, instance, **kwargs):
     global post_delete_flag
     post_delete_flag = True
 
-    post_media = instance.postmedia_set.all()
-
-    if post_media.exists():
-        media_dir = os.path.dirname(post_media.first().file.path)
-
-        for media in post_media:
-            media.file.delete()
-
-        if os.path.exists(media_dir):
-            os.rmdir(media_dir)
+    for media in instance._media_files:
+        media.file.delete()
+    
+    if hasattr(instance, '_media_dir') and os.path.exists(instance._media_dir):
+        os.rmdir(instance._media_dir)
     
     post_delete_flag = False
